// src/types/App/theme.d.ts
export type ThemeName = 'light' | 'dark'

export interface ThemeSettings {
    black: string
    backgroundContent: string
    borderColor: string
    titleColor: string
    subtitleColor: string
    tabBorderColor: string
    selectColor: string
    inputPlaceholderColor: string
    navbarBg: string
    inputWrapperColor: string
    inputBorderColor: string
    inputDisabledBorderColor: string
    inputWrapperDisabledColor: string
    arrowBackgroundColor: string
    faqColor: string
    faqBorderColor: string
    faqCrossColor: string
    homePageButton: string
    navbarBottom: string
    walletBorder: string
    navbarHover: string
    navbarHoverFill: string
}